requirements.txt
openai 
python-dotenv 
crewai 
langchain 
pandas 
plotly
protobuf
streamlit
graphviz